refactor: update type imports and extensions for consistency across modules
