+++ src/components/RelayEditor.vue
@@ -4,7 +4,7 @@ import { useQuasar } from 'quasar';
 import { useI18n } from 'vue-i18n';
 import { normalizeRelayUrl } from 'src/services/relay-url';
 import type { NostrRelay, StoredKey } from '../types';
-import type { RelayCatalogEntry } from 'src/types/relay';
+import type { RelayCatalogEntry } from '../types/relay';
 import { SimplePool } from 'nostr-tools';
 import { finalizeEvent } from 'nostr-tools/pure';
 import { hexToBytes } from '@noble/hashes/utils';
